Began moving markup code out.
