+++ src/antsibull_docs/markup/rstify.py
@@ -3,26 +3,46 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # SPDX-FileCopyrightText: 2022, Ansible Project
 """
-htmlify Jinja2 filter for use in Ansible documentation.
+rstify Jinja2 filter for use in Ansible documentation.
 """
 
 import re
 import typing as t
-from html import escape as html_escape
 from urllib.parse import quote
 
-from antsibull_core.logging import log
-from jinja2.runtime import Context
-from jinja2.utils import pass_context
-
-from ..semantic_helper import parse_option, parse_return_value
-from .filters import extract_plugin_data
+from .semantic_helper import augment_plugin_name_type
 from .parser import Command, CommandSet, convert_text
-
-mlog = log.fields(mod=__name__)
 
 _MODULE = re.compile(r"^([^).]+)\.([^).]+)\.([^)]+)$")
 _PLUGIN = re.compile(r"^([^).]+)\.([^).]+)\.([^)]+)#([a-z]+)$")
+
+
+def rst_escape(value: t.Any, escape_ending_whitespace=False) -> str:
+    ''' make sure value is converted to a string, and RST special characters are escaped '''
+
+    if not isinstance(value, str):
+        value = str(value)
+
+    value = value.replace('\\', '\\\\')
+    value = value.replace('<', '\\<')
+    value = value.replace('>', '\\>')
+    value = value.replace('_', '\\_')
+    value = value.replace('*', '\\*')
+    value = value.replace('`', '\\`')
+
+    if escape_ending_whitespace and value.endswith(' '):
+        value = value + '\\ '
+    if escape_ending_whitespace and value.startswith(' '):
+        value = '\\ ' + value
+
+    return value
+
+
+def rst_code(value: str) -> str:
+    ''' Write value as :code:`...` RST construct. '''
+    if not isinstance(value, str):
+        value = str(value)
+    return f':code:`{rst_escape(value, escape_ending_whitespace=True)}`'
 
 
 def _escape_url(url: str) -> str:
@@ -32,20 +52,19 @@ def _escape_url(url: str) -> str:
 
 
 def _create_error(text: str, error: str) -> str:
-    text = f'<code>{html_escape(text)}</code>'
-    return f'<span class="error">ERROR while parsing {text}: {html_escape(error)}</span>'
+    text = f':literal:`{rst_escape(text, escape_ending_whitespace=True)}`'
+    error_msg = f':strong:`{rst_escape(error, escape_ending_whitespace=True)}`'
+    return f"\\ :strong:`ERROR while parsing` {text}\\ : {error_msg}\\ "
 
 
 class _Context:
-    j2_context: Context
     counts: t.Dict[str, int]
     plugin_fqcn: t.Optional[str]
     plugin_type: t.Optional[str]
 
-    def __init__(self, j2_context: Context,
+    def __init__(self,
                  plugin_fqcn: t.Optional[str] = None,
                  plugin_type: t.Optional[str] = None):
-        self.j2_context = j2_context
         self.counts = {
             'italic': 0,
             'bold': 0,
@@ -61,8 +80,8 @@ class _Context:
             'return-value': 0,
             'ruler': 0,
         }
-        self.plugin_fqcn, self.plugin_type = extract_plugin_data(
-            j2_context, plugin_fqcn=plugin_fqcn, plugin_type=plugin_type)
+        self.plugin_fqcn = plugin_fqcn
+        self.plugin_type = plugin_type
 
 
 # In the following, we make heavy use of escaped whitespace ("\ ") being removed from the output.
@@ -78,7 +97,7 @@ class _Italic(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['italic'] += 1
-        return f"<em>{html_escape(parameters[0])}</em>"
+        return f"\\ :emphasis:`{rst_escape(parameters[0], escape_ending_whitespace=True)}`\\ "
 
 
 class _Bold(Command):
@@ -88,7 +107,7 @@ class _Bold(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['bold'] += 1
-        return f"<b>{html_escape(parameters[0])}</b>"
+        return f"\\ :strong:`{rst_escape(parameters[0], escape_ending_whitespace=True)}`\\ "
 
 
 class _Module(Command):
@@ -103,8 +122,7 @@ class _Module(Command):
             return _create_error(
                 f'M({parameters[0]!r})', f'parameter {parameters[0]!r} is not a FQCN')
         fqcn = f'{m.group(1)}.{m.group(2)}.{m.group(3)}'
-        url = html_escape(f'../../{m.group(1)}/{m.group(2)}/{m.group(3)}_module.html')
-        return f"<a href='{url}' class='module'>{html_escape(fqcn)}</a>"
+        return f"\\ :ref:`{rst_escape(fqcn)} <ansible_collections.{fqcn}_module>`\\ "
 
 
 class _Plugin(Command):
@@ -121,9 +139,7 @@ class _Plugin(Command):
                 f'parameter {parameters[0]!r} is not of the form FQCN#type')
         fqcn = f'{m.group(1)}.{m.group(2)}.{m.group(3)}'
         plugin_type = m.group(4)
-        url = html_escape(f'../../{m.group(1)}/{m.group(2)}/{m.group(3)}_{plugin_type}.html')
-        cssclass = f'plugin-{html_escape(plugin_type)}'
-        return f"<a href='{url}' class='module {cssclass}'>{html_escape(fqcn)}</a>"
+        return f"\\ :ref:`{rst_escape(fqcn)} <ansible_collections.{fqcn}_{plugin_type}>`\\ "
 
 
 class _URL(Command):
@@ -133,8 +149,7 @@ class _URL(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['url'] += 1
-        url = parameters[0]
-        return f"<a href='{html_escape(_escape_url(url))}'>{html_escape(_escape_url(url))}</a>"
+        return f"\\ {_escape_url(parameters[0])}\\ "
 
 
 class _Link(Command):
@@ -144,8 +159,7 @@ class _Link(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['link'] += 1
-        url = parameters[1]
-        return f"<a href='{html_escape(_escape_url(url))}'>{html_escape(parameters[0])}</a>"
+        return f"\\ `{rst_escape(parameters[0])} <{_escape_url(parameters[1])}>`__\\ "
 
 
 class _Ref(Command):
@@ -155,7 +169,7 @@ class _Ref(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['ref'] += 1
-        return f"<span class='module'>{html_escape(parameters[0])}</span>"
+        return f"\\ :ref:`{rst_escape(parameters[0])} <{parameters[1]}>`\\ "
 
 
 class _Const(Command):
@@ -166,7 +180,7 @@ class _Const(Command):
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['const'] += 1
         # Escaping does not work in double backticks, so we use the :literal: role instead
-        return f"<code class='docutils literal notranslate'>{html_escape(parameters[0])}</code>"
+        return f"\\ :literal:`{rst_escape(parameters[0], escape_ending_whitespace=True)}`\\ "
 
 
 class _OptionName(Command):
@@ -178,39 +192,8 @@ class _OptionName(Command):
         context.counts['option-name'] += 1
         if context.plugin_fqcn is None or context.plugin_type is None:
             raise Exception('The markup O(...) cannot be used outside a plugin or role')
-        text = parameters[0]
-        try:
-            plugin_fqcn, plugin_type, option_link, option, value = parse_option(
-                text, context.plugin_fqcn, context.plugin_type, require_plugin=False)
-        except ValueError as exc:
-            return _create_error(f'O({text})', str(exc))
-        if value is None:
-            cls = 'ansible-option'
-            text = f'{option}'
-            strong_start = '<strong>'
-            strong_end = '</strong>'
-        else:
-            cls = 'ansible-option-value'
-            text = f'{option}={value}'
-            strong_start = ''
-            strong_end = ''
-        if plugin_fqcn and plugin_type and plugin_fqcn.count('.') >= 2:
-            # TODO: handle role arguments (entrypoint!)
-            namespace, name, plugin = plugin_fqcn.split('.', 2)
-            url = f'../../{namespace}/{name}/{plugin}_{plugin_type}.html'
-            fragment = f'parameter-{quote(option_link.replace(".", "/"))}'
-            link_start = (
-                f'<a class="reference internal" href="{url}#{fragment}">'
-                '<span class="std std-ref"><span class="pre">'
-            )
-            link_end = '</span></span></a>'
-        else:
-            link_start = ''
-            link_end = ''
-        return (
-            f'<code class="{cls} literal notranslate">'
-            f'{strong_start}{link_start}{html_escape(text)}{link_end}{strong_end}</code>'
-        )
+        text = augment_plugin_name_type(parameters[0], context.plugin_fqcn, context.plugin_type)
+        return f"\\ :ansopt:`{rst_escape(text, escape_ending_whitespace=True)}`\\ "
 
 
 class _OptionValue(Command):
@@ -220,8 +203,7 @@ class _OptionValue(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['option-value'] += 1
-        text = parameters[0]
-        return f'<code class="ansible-value literal notranslate">{html_escape(text)}</code>'
+        return f"\\ :ansval:`{rst_escape(parameters[0], escape_ending_whitespace=True)}`\\ "
 
 
 class _EnvVariable(Command):
@@ -231,8 +213,7 @@ class _EnvVariable(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['environment-var'] += 1
-        text = parameters[0]
-        return f'<code class="xref std std-envvar literal notranslate">{html_escape(text)}</code>'
+        return f"\\ :envvar:`{rst_escape(parameters[0], escape_ending_whitespace=True)}`\\ "
 
 
 class _RetValue(Command):
@@ -244,33 +225,8 @@ class _RetValue(Command):
         context.counts['return-value'] += 1
         if context.plugin_fqcn is None or context.plugin_type is None:
             raise Exception('The markup RV(...) cannot be used outside a plugin or role')
-        text = parameters[0]
-        try:
-            plugin_fqcn, plugin_type, rv_link, rv, value = parse_return_value(
-                text, context.plugin_fqcn, context.plugin_type, require_plugin=False)
-        except ValueError as exc:
-            return _create_error(f'RV({text})', str(exc))
-        cls = 'ansible-return-value'
-        if value is None:
-            text = f'{rv}'
-        else:
-            text = f'{rv}={value}'
-        if plugin_fqcn and plugin_type and plugin_fqcn.count('.') >= 2:
-            namespace, name, plugin = plugin_fqcn.split('.', 2)
-            url = f'../../{namespace}/{name}/{plugin}_{plugin_type}.html'
-            fragment = f'return-{quote(rv_link.replace(".", "/"))}'
-            link_start = (
-                f'<a class="reference internal" href="{url}#{fragment}">'
-                '<span class="std std-ref"><span class="pre">'
-            )
-            link_end = '</span></span></a>'
-        else:
-            link_start = ''
-            link_end = ''
-        return (
-            f'<code class="{cls} literal notranslate">{link_start}'
-            f'{html_escape(text)}{link_end}</code>'
-        )
+        text = augment_plugin_name_type(parameters[0], context.plugin_fqcn, context.plugin_type)
+        return f"\\ :ansretval:`{rst_escape(text, escape_ending_whitespace=True)}`\\ "
 
 
 class _HorizontalLine(Command):
@@ -280,7 +236,7 @@ class _HorizontalLine(Command):
 
     def handle(self, parameters: t.List[str], context: t.Any) -> str:
         context.counts['ruler'] += 1
-        return '<hr/>'
+        return '\n\n.. raw:: html\n\n  <hr>\n\n'
 
 
 _COMMAND_SET = CommandSet([
@@ -300,26 +256,19 @@ _COMMAND_SET = CommandSet([
 ])
 
 
-@pass_context
-def html_ify(context: Context, text: str,
-             *,
-             plugin_fqcn: t.Optional[str] = None,
-             plugin_type: t.Optional[str] = None) -> str:
-    ''' convert symbols like I(this is in italics) to valid HTML '''
-    flog = mlog.fields(func='html_ify')
-    flog.fields(text=text).debug('Enter')
-
+def rst_ify(text: str,
+            *,
+            plugin_fqcn: t.Optional[str] = None,
+            plugin_type: t.Optional[str] = None) -> t.Tuple[str, t.Mapping[str, int]]:
+    ''' convert symbols like I(this is in italics) to valid restructured text '''
     our_context = _Context(
-        context,
         plugin_fqcn=plugin_fqcn,
         plugin_type=plugin_type,
     )
 
     try:
-        text = convert_text(text, _COMMAND_SET, html_escape, our_context)
+        text = convert_text(text, _COMMAND_SET, rst_escape, our_context)
     except Exception as exc:  # pylint:disable=broad-except
-        return _create_error(text, str(exc))
+        text = _create_error(text, str(exc))
 
-    flog.fields(counts=our_context.counts).info('Number of macros converted to html equivalents')
-    flog.debug('Leave')
-    return text
+    return text, our_context.counts
